requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
decouple==0.0.7
Django==4.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
idna==3.11
numpy==2.3.5
pandas==2.3.3
phonenumbers==8.13.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
